+++ source/internal_extensions/isaacsim.util.debug_draw/nodes/OgnDebugDrawRayCast.cpp
@@ -116,6 +116,6 @@ private:
 };
 
 REGISTER_OGN_NODE()
-} // debug_draw
-} // isaac
-} // omni
+}
+}
+}